API Reference

PoGO.jl exported functions

PoGO.approximate — Function

approximate(
    x::Union{VariableRef,AffExpr},
    func::Function,
    n::Int = 0;
    δ::Real = 0.0,
    method::Symbol = :default,
    type::Symbol = :interior,
    knots::Union{Nothing,Vector{Float64},Vector{Tuple{Float64,Symbol}}} = nothing,
    name::String = "",
)

Function that defines a mixed-integer programming formulation for an arbitrary univariate function.

Required arguments

x is either of type VariableRef or AffExpr and is the input variable for the function func.

func is the function that we are approximating.

Optional arguments

n the number of points sampled between knots (ignored if δ is used).

δ is the minimum distance between sample points.

method is the method of the approximation. This can be set to :binary, :echelon, :convex, :bisection, :SOS1 or :SOS2.

type specifies how the approximation bounds the actual function. This can be set to :upper, :lower, :interior, tangent_cuts or :combined.

knots must list the function's discontinuities, and points of inflection. Can be provided with curvature specifed, but if not this will be determined automatically.

name can be set to give the variables created meaningful names.

PoGO.bilinear — Function

bilinear(
    x::Union{VariableRef,AffExpr},
    y::Union{VariableRef,AffExpr},
    n::Int = 0;
    method::Symbol = :default,
    type::Symbol = :interior,
    name::String = "",
)

Function that defines a bilinear formulation for a given pair of variables or affine expressions. This function will select the best formulation, given x and y.

Required arguments

x and y are either of type VariableRef or AffExpr and xy is the product that is being modelled.

Optional arguments

n is used if the product is approximated using the difference of two squares.

method is used if the product is approximated using the difference of two squares. This can be set to :binary, :echelon, :SOS1 or :SOS2.

type is used if the product is approximated using the difference of two squares. This can be set to :upper, :lower, :interior, cutting_planes or :combined.

name can be set to give the variables created meaningful names.

PoGO.interpolate — Function

interpolate(
    x_vector::Vector{<:Union{VariableRef,AffExpr}},
    points::Vector{<:Tuple},
    sets::Vector;
    update_bounds::Bool = false,
    method::Symbol = :default,
    name::String = "",
)

Function that constrains the variables to lie inside the convex hull of one of the sets of points (defined by sets).

Required arguments

x_vector is the vector of variables or expressions that will be used as the domain of the interpolation

points is the vector of points that define the interpolation; these points can contain real number as well as variables and expressions.

sets specifies the name(s) of the set(s) of which each point in the vector points is a member.

Optional arguments

update_bounds is set to true if the upper and lower bounds of x_vector components should be updated based on the values in points.

method is the formulation method and can be set to :convex, :SOS1, :binary or :bisection.

name can be set to give the variables created meaningful names.

PoGO.interpolate_fn — Function

interpolate_fn(
    f::Function,
    x_vector::Vector{<:Union{VariableRef,AffExpr}},
    grid::Vector{<:Union{AbstractRange,Vector{<:Real}}};
    method::Symbol = :default,
)

Function that interpolates a function f over a set of variables (or affine expressions). The grid vector should be the same length as the x_vector.

Required arguments

f a possibly multidimensional function that takes a vector of arguments the same length as x_vector.

x_vector a vector of variables or expressions that are the input vector for f.

grid is a vector of ranges or vectors that specify the sample points (which are the Cartesian product of the vectors) for the function f.

Optional arguments

method if the formulation method and can be set to :convex, :SOS1, :binary or :bisection.

interpolate_fn(
    f::Function,
    x_vector::Vector{<:Union{VariableRef,AffExpr}},
    points::T where {T<:Matrix};
    method::Symbol = :default,
)

Function that interpolates a function f over a set of variables (or affine expressions). The number of columns of the points matrix should be the same as the length of the x_vector.

Required arguments

f is a possibly multidimensional function that takes a vector of arguments the same length as x_vector.

x_vector is a vector of variables or expressions that are the input vector for f.

points is a Matrix that specifies the sample points for the function.

Optional arguments

method if the formulation method and can be set to :convex, :SOS1, :binary or :bisection.

PoGO.interpolate_points — Function

interpolate_points(
    x_vector::Vector{<:Union{VariableRef,AffExpr}},
    points::T where {T<:Matrix};
    method::Symbol = :default,
)

Function that interpolates over a set of variables (or affine expressions). The number of columns of the points matrix should be greater than or equal to the length of the x_vector. For any extra columns, new variables will be defined, and their values will be interpolated.

Required arguments

x_vector is a vector of variables or expressions around which the triangulation will be formed.

points is a Matrix that specifies the points to be interpolated.

Optional arguments

method is the formulation method and can be set to :convex, :SOS1, :binary or :bisection.

PoGO.plot_approximation — Function

plot_approximation(
    optimizer,
    func::Function,
    lb::Real,
    ub::Real,
    δ::Real,
    type::Symbol;
    detail::Int = 3,
    knots::Union{Nothing,Vector{Float64},Vector{Tuple{Float64,Symbol}}} = nothing,
)

Function that returns a set of points corresponding to an approximation of the function func.

PoGO.power — Function

power(
    x::Union{VariableRef,AffExpr},
    y::Union{VariableRef,AffExpr},
    n::Int = 0;
    method::Symbol = :default,
    type::Symbol = :interior,
    name::String = "",
)

Function that defines a power formulation (x^y) for a given pair of variables or affine expressions. This function will select the best formulation, given x and y.

Required arguments

x and y are either of type VariableRef or AffExpr and x^y is the power function that is being modelled.

Optional arguments

n is used if the power is approximated using the difference of two squares.

method is used if the power is approximated using the difference of two squares. This can be set to :binary :echelon :SOS1 or :SOS2.

type is used if the power is approximated using the difference of two squares. This can be set to :upper :lower :interior cutting_planes or :combined.

name can be set to give the variables created meaningful names.

PoGO.set_var_domain — Function

set_var_domain(x::Union{VariableRef,AffExpr}, domain::Vector)

Function that specifies a domain for variable x. This can include discrete values as well as intervals.

Required arguments

x is either a VariableRef or an AffExpr.

domain is a vector of values and / or intervals, defining the feasible region for x.

PoGO.jl internal functions

PoGO.one_binary_bilinear_formulation — Function

one_binary_bilinear_formulation(
    model::JuMP.Model,
    xy::VariableRef,
    x::Union{VariableRef,AffExpr},
    y::Union{VariableRef,AffExpr},
)

Internal function that defines a bilinear formulation when one of the variables is binary.

PoGO.discrete_bilinear_formulation — Function

discrete_bilinear_formulation(
    model::JuMP.Model,
    xy::VariableRef,
    x::Union{VariableRef,AffExpr},
    y::Union{VariableRef,AffExpr},
    x_values::Vector,
    mx::Float64,
    mn::Float64,
)

Internal function that defines a bilinear formulation when one of the variables is discrete.

PoGO.find_bounds — Function

find_bounds(x::Union{VariableRef,AffExpr}; ignore_errors = false)

Internal function that finds the upper and lower bounds of a variable or affine expression. ignore_errors can be set to true if we are happy to get Inf or -Inf if the bounds are not defined.

PoGO.find_points — Function

find_points(
    func::Function,
    n::Integer,
    δ::Real,
    knots::Vector{Float64},
    knots_shape::Vector{Symbol},
    type::Symbol,
)

Internal function that finds all the points and corresponding function values, given knots and the sampling rate.

PoGO.get_model — Function

get_model(x::Union{VariableRef,AffExpr})

Internal function that gets the model associated with a variable or affine expression.

PoGO.get_type — Function

get_type(x::Union{VariableRef,AffExpr})

Internal function that gets the type of a variable or affine expression. This can be :binary, :integer, :continuous or :discrete.

PoGO.infer_curvature — Function

infer_curvature(
    f::Function,
    knots::Union{Nothing,Vector{Float64},Vector{Tuple{Float64,Symbol}}},
    lx::Real,
    ux::Real,
)

Internal function that uses automatic differentiation to infer the curvature of the function f between the bounds and the knots.

PoGO.process_knots — Function

process_knots(
    knots::Vector{Tuple{Float64,Symbol}},
    lx::Real,
    ux::Real,
)::Tuple{Vector{Float64},Vector{Symbol}}

Internal function that trims or extends the knots, given the lower and upper bounds.

PoGO.setup_parametric_model — Function

setup_parametric_model(
    optimizer,
    func::Function,
    knots::Union{Nothing,Vector{Float64},Vector{Tuple{Float64,Symbol}}},
    lb::Real,
    ub::Real,
    δ::Real,
    type::Symbol,
)

Internal function that creates a simple JuMP.Model to evaluate an interpolated function.